gofish_project/requirements.txt
django==1.5.4
mock==1.0.1
django-lazysignup==0.12.2
python-constraint==1.2.0
